server/requirements.txt
Authlib==0.11
bandit==1.5.1
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
docutils==0.15.2
flake8==3.6.0
Flask==1.0.2
gitdb2==2.0.6
GitPython==3.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
pbr==5.4.3
pkginfo==1.5.0.1
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.4.2
python-dotenv==0.10.3
PyYAML==5.1.2
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
smmap2==2.0.5
stevedore==1.31.0
tqdm==4.36.1
twine==1.13.0
urllib3==1.24.2
webencodings==0.5.1
Werkzeug==0.16.0
